requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
requests==2.31.0
python-decouple==3.8
django-filter==23.3
drf-spectacular==0.27.0
gunicorn==21.2.0
mysqlclient==2.2.4 